feat(insights): add copy as CSV to the dashboard insight card menu
Adds a client-side "Copy as CSV" option to the insight card more menu for trends insights, so loaded results can go straight to the clipboard without a backend export. The generic insight export adapters move out of web analytics into a shared module, and the trends adapter learns to handle value-aggregated displays.

Generated-By: PostHog Code
Task-Id: 2df25632-4ae1-4e6c-9f8b-75a69424f33f
